get search

+++ test/setup.ts
@@ -3,24 +3,26 @@ import { PrismaClient } from "../src/generated/prisma";
 
 const prisma = new PrismaClient();
 
-beforeAll(() => {
-  execSync("npx prisma db push --force-reset");
-});
+if (process.env.TEST_TYPE === "integration") {
+  beforeAll(() => {
+    execSync("npx prisma db push --force-reset");
+  });
 
-beforeEach(async () => {
-  const tablenames = await prisma.$queryRaw<
-    Array<{ tablename: string }>
-  >`SELECT tablename FROM pg_tables WHERE schemaname='public'`;
+  beforeEach(async () => {
+    const tablenames = await prisma.$queryRaw<
+      Array<{ tablename: string }>
+    >`SELECT tablename FROM pg_tables WHERE schemaname='public'`;
 
-  const tables = tablenames
-    .map(({ tablename }) => tablename)
-    .filter((name) => name !== "_prisma_migrations")
-    .map((name) => `"public"."${name}"`)
-    .join(", ");
+    const tables = tablenames
+      .map(({ tablename }) => tablename)
+      .filter((name) => name !== "_prisma_migrations")
+      .map((name) => `"public"."${name}"`)
+      .join(", ");
 
-  try {
-    await prisma.$executeRawUnsafe(`TRUNCATE TABLE ${tables} CASCADE;`);
-  } catch (error) {
-    console.log({ error });
-  }
-});
+    try {
+      await prisma.$executeRawUnsafe(`TRUNCATE TABLE ${tables} CASCADE;`);
+    } catch (error) {
+      console.log({ error });
+    }
+  });
+}
